shell script parse bug fixed.
